rename telemetry_events to telemetry
