+++ crates/assistant/Cargo.toml
@@ -77,7 +77,7 @@ similar.workspace = true
 smallvec.workspace = true
 smol.workspace = true
 strum.workspace = true
-telemetry_events.workspace = true
+telemetry.workspace = true
 terminal.workspace = true
 terminal_view.workspace = true
 text.workspace = true
