+++ crates/client/Cargo.toml
@@ -42,7 +42,7 @@ serde_json.workspace = true
 settings.workspace = true
 sha2.workspace = true
 smol.workspace = true
-telemetry_events.workspace = true
+telemetry.workspace = true
 text.workspace = true
 thiserror.workspace = true
 time.workspace = true
