+++ crates/remote_server/Cargo.toml
@@ -54,7 +54,7 @@ settings.workspace = true
 shellexpand.workspace = true
 smol.workspace = true
 sysinfo.workspace = true
-telemetry_events.workspace = true
+telemetry.workspace = true
 util.workspace = true
 worktree.workspace = true
