+++ crates/telemetry/LICENSE-GPL
