+++ crates/zed/Cargo.toml
@@ -103,7 +103,7 @@ sysinfo.workspace = true
 tab_switcher.workspace = true
 task.workspace = true
 tasks_ui.workspace = true
-telemetry_events.workspace = true
+telemetry.workspace = true
 terminal_view.workspace = true
 theme.workspace = true
 theme_selector.workspace = true
